Move PMA managed-thread runtime orchestration into services
